Add videos to slider
