+++ data/generateProjectData.ts
@@ -10,6 +10,7 @@ interface GenerateProjectDataArgs {
     thumbnailImageName: string
     metaImageName?: string
     imagesNames?: string[]
+    videosNames?: string[]
     date: string
     youtubeVideoIds?: string[]
     previewImageName?: string
@@ -34,6 +35,7 @@ const generateProjectData = ({
     thumbnailImageName,
     metaImageName,
     imagesNames = [],
+    videosNames = [],
     youtubeVideoIds = [],
     previewImageName = null,
     title,
@@ -54,6 +56,7 @@ const generateProjectData = ({
     },
 }: GenerateProjectDataArgs): ProjectData => {
     const imagesFolder = `/images/portfolio/${id}`
+    const videosFolder = `/videos/${id}`
 
     const data: ProjectData = {
         id,
@@ -62,6 +65,7 @@ const generateProjectData = ({
         thumbnailUrl: `${imagesFolder}/${thumbnailImageName}`,
         metaImageUrl: `${imagesFolder}/${metaImageName || thumbnailImageName}`,
         imagesUrls: imagesNames.map((name) => `${imagesFolder}/${name}`),
+        videosUrls: videosNames.map((name) => `${videosFolder}/${name}`),
         youtubeVideoIds,
         previewImageUrl: previewImageName
             ? `${imagesFolder}/${previewImageName}`
