+++ data/projects/cg-ball-of-lava-shader-unity.ts
@@ -8,14 +8,14 @@ const data: ProjectData = generateProjectData({
     aliases: ['ball-of-lava-shader-unity'],
     date: '2022-10',
     thumbnailImageName: 'thumbnail.gif',
-    imagesNames: [
-        '7.gif',
-        '6.gif',
-        '5.gif',
-        '4.gif',
-        '3.gif',
-        '2.gif',
-        '1.gif',
+    videosNames: [
+        '7.mp4',
+        '6.mp4',
+        '5.mp4',
+        '4.mp4',
+        '3.mp4',
+        '2.mp4',
+        '1.mp4',
     ],
     youtubeVideoIds: ['r5qDHvfYG9c'],
     title: 'Ball of Lava Shader',
@@ -37,8 +37,8 @@ const data: ProjectData = generateProjectData({
     ],
     technology: Technology.UnityEngine,
     cardConfig: {
-        backgroundSize: '120%',
-        backgroundPosition: '13px 3px',
+        backgroundSize: '100%',
+        backgroundPosition: '22px -38px',
         gradientColor: 'rgb(109, 23, 0)',
     },
 })
