+++ src/models/enums/Technology.ts
@@ -1,3 +1,4 @@
+/* eslint-disable unused-imports/no-unused-vars */
 export enum Technology {
     UnityEngine = 'UnityEngine',
     UnrealEngine = 'UnrealEngine',
